third_party/blink/web_tests/fast/events/onunload-clears-onbeforeunload-expected.txt

CONSOLE ERROR: Blocked alert('before unload') during beforeunload.
CONSOLE ERROR: Blocked alert('unload') during unload.
If you didn't hit an assert you PASS.
